pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ebird-recommend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.0",
    "typer>=0.12",
    "rich>=13.0",
    "python-dotenv>=1.0",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
]

[project.scripts]
ebird-rec = "ebird_recommend.cli.app:app"

[tool.hatch.build.targets.wheel]
packages = ["ebird_recommend"]
